# openapi.yaml
openapi: 3.1.0
info:
  title: PM Assistant API
  version: 1.0.0
  description: >
    REST API สำหรับผู้ช่วย PM บน LINE — ใช้โดยหน้าเว็บตั้งค่า (LIFF) และรับ event จาก
    LINE/แพลตฟอร์มภายนอก. ใช้คู่กับ schema.sql และ DESIGN_SPEC.md.
    ทุก endpoint ฝั่ง LIFF ต้องมี bearer token (จาก LINE Login); endpoint /webhooks
    ตรวจ signature แทน.
servers:
  - url: https://api.pmassistant.app/v1
    description: production
  - url: http://localhost:8000/v1
    description: local dev
...
security:
  - bearerAuth: []
paths:
...
  /webhooks/line:
    post:
      tags: [Webhooks]
      summary: รับ event จาก LINE
      description: ตรวจ X-Line-Signature ก่อนประมวลผลเสมอ ไม่ใช้ bearer token
      security: []
      parameters:
        - name: X-Line-Signature
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                destination: { type: string }
                events:
                  type: array
                  items: { type: object, additionalProperties: true }
      responses:
        '200': { description: รับแล้ว (ประมวลผลแบบ async) }
        '403': { $ref: '#/components/responses/Forbidden' }
  /webhooks/{provider}:
    post:
      tags: [Webhooks]
      summary: รับ webhook จาก tracker (เช่นสถานะ Jira/ClickUp เปลี่ยน)
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, enum: [jira, clickup] }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: รับแล้ว }

  # ---------------------------------------------------------------- Projects
...
  /projects/{projectId}/traceability:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [Traceability]
      summary: ความครอบคลุม req ↔ task ↔ test (หน้า W9)
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TraceabilityCoverage' }
  /projects/{projectId}/traceability/links:
    parameters:
      - $ref: '#/components/parameters/projectId'
    post:
      tags: [Traceability]
      summary: ผูก requirement เข้ากับ task หรือ test case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requirementId]
              properties:
                requirementId: { type: string, format: uuid }
                taskId: { type: string, format: uuid, nullable: true }
                testCaseId: { type: string, format: uuid, nullable: true }
      responses:
        '201': { description: สำเร็จ }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ----------------------------------------------------------- Confirmations
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: token ที่ออกหลังตรวจ LINE Login ID token จากฝั่ง LIFF
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Forbidden:
      description: สิทธิ์ไม่พอ / signature ไม่ถูกต้อง
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: ข้อมูลไม่ผ่านการตรวจสอบ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: not_found }
        message: { type: string }
        details: { type: object, additionalProperties: true }

    # ---- projects ----
...
    TraceabilityCoverage:
      type: object
      properties:
        requirementId: { type: string, format: uuid }
        code: { type: string, example: REQ-012 }
        title: { type: string }
        coverage: { type: string, enum: [covered, missing_test, no_task] }
        taskKeys:
          type: array
          items: { type: string }
        testCaseCodes:
          type: array
          items: { type: string }

    # ---- confirmations ----
...